Added static object (STAT) record support Added YAML output for static objects, interior cells, their persistent objects, and worldspaces (including their cells and their cell's objects)

+++ source/ESM/ESMStream.hpp
@@ -81,7 +81,7 @@ namespace ESM
             mOffset += length;
         }
         
-        inline void ReadCString(uint16_t length, std::string& outValue) {
+        inline void ReadCString(size_t length, std::string& outValue) {
             if (mOffset + length <= mLocalSize) {
                 mBuffer.reserve(length);
                 mStream.read((char *)&mBuffer[0], length);
